Major refactor complete
