+++ cppapplication/README.md
@@ -1,12 +1,14 @@
 # cuda_project c++ implementation
 
-# Steps to build and run
+## Steps to build and run
 
-## Install docker
-https://docs.docker.com/desktop/
+### Install docker
 
-# Run commands in the cppapplication directory 
+<https://docs.docker.com/desktop/>
+
+### Run commands in the cppapplication directory
+
 ```bash
-docker build -t ascii_rt .
-docker run ascii_rt
+sudo docker build -t ascii_rt .
+sudo docker run -i ascii_rt < ../models/teapot.obj
 ```
