+++ cppapplication/src/utils/intersect.hpp
@@ -1,18 +1,15 @@
-#ifndef INTERSECT_VISITOR_HPP
-#define INTERSECT_VISITOR_HPP
-
+#ifndef INTERSECT_HPP
+#define INTERSECT_HPP
+#include "util.hpp"
 #include "ray.hpp"
-#include "triangle.hpp"
-#include "sphere.hpp"
-#include "mesh.hpp"
-#include <optional>
-#include <tuple>
-#include <Eigen/Dense>
+#include "../shapes/triangle.hpp"
+#include "../shapes/sphere.hpp"
+#include "../shapes/mesh.hpp"
 
-// IntersectVisitor class definition
-class IntersectVisitor {
+
+class Intersect {
 public:
-    IntersectVisitor(const Ray& ray);
+    Intersect(const Ray& ray);
     std::optional<std::tuple<double, Eigen::Vector3d>> operator()(const Triangle& triangle) const;
     std::optional<std::tuple<double, Eigen::Vector3d>> operator()(const Sphere& sphere) const;
     std::optional<std::tuple<double, Eigen::Vector3d>> operator()(const Mesh& mesh) const;
@@ -21,4 +18,4 @@ private:
     const Ray& ray;
 };
 
-#endif // INTERSECT_VISITOR_HPP
+#endif // INTERSECT_HPP
