+++ cppapplication/src/utils/ray.cpp
