+++ cppapplication/src/utils/ray.hpp
