+++ cppapplication/src/utils/util.hpp
